+++ src/Execution/DataLoader/ModelRelationFetcher.php
@@ -19,7 +19,7 @@ class ModelRelationFetcher
     /**
      * The parent models that relations should be loaded for.
      *
-     * @var \Illuminate\Database\Eloquent\Collection
+     * @var \Illuminate\Database\Eloquent\Collection<\Illuminate\Database\Eloquent\Model>
      */
     protected $models;
 
@@ -45,6 +45,8 @@ class ModelRelationFetcher
 
     /**
      * Load all relations for the model, but constrain the query to the current page.
+     *
+     * @return \Illuminate\Database\Eloquent\Collection<\Illuminate\Database\Eloquent\Model>
      */
     public function loadRelationsForPage(PaginationArgs $paginationArgs): EloquentCollection
     {
@@ -57,12 +59,15 @@ class ModelRelationFetcher
 
     /**
      * Reload the models to get the `{relation}_count` attributes of models set.
+     *
+     * @return \Illuminate\Database\Eloquent\Collection<\Illuminate\Database\Eloquent\Model>
      */
     public function reloadModelsWithRelationCount(): EloquentCollection
     {
         $ids = $this->models->modelKeys();
 
-        $this->models = $this
+        /** @var \Illuminate\Database\Eloquent\Collection<\Illuminate\Database\Eloquent\Model> $reloadedModels */
+        $reloadedModels = $this
             ->newModelQuery()
             ->withCount($this->relations)
             ->whereKey($ids)
@@ -77,7 +82,7 @@ class ModelRelationFetcher
                 );
             });
 
-        return $this->models;
+        return $this->models = $reloadedModels;
     }
 
     /**
@@ -208,7 +213,7 @@ class ModelRelationFetcher
             ->reduce(
                 function (EloquentBuilder $builder, Relation $relation) {
                     return $builder->unionAll(
-                        $relation->getQuery()
+                        $relation->getBaseQuery()
                     );
                 },
                 // Use the first query as the initial starting point