+++ src/Schema/Types/Scalars/DateScalar.php
@@ -49,7 +49,7 @@ abstract class DateScalar extends ScalarType
         if (! $valueNode instanceof StringValueNode) {
             throw new Error(
                 "Query error: Can only parse strings, got {$valueNode->kind}",
-                [$valueNode]
+                $valueNode
             );
         }
 
@@ -60,7 +60,7 @@ abstract class DateScalar extends ScalarType
      * Try to parse the given value into a Carbon instance, throw if it does not work.
      *
      * @param  string  $value
-     * @param  string|\Exception  $exceptionClass
+     * @param  class-string<\Exception>  $exceptionClass
      *
      * @throws \GraphQL\Error\InvariantViolation|\GraphQL\Error\Error
      */