+++ src/WhereConditions/WhereConditionsDirective.php
@@ -28,11 +28,6 @@ directive @whereConditions(
 SDL;
     }
 
-    /**
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @param  mixed[]  $whereConditions
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
-     */
     public function handleBuilder($builder, $whereConditions)
     {
         // The value `null` should be allowed but have no effect on the query.