+++ tests/Unit/Subscriptions/SubscriberTest.php
@@ -2,6 +2,8 @@
 
 namespace Tests\Unit\Subscriptions;
 
+use GraphQL\Language\AST\NameNode;
+use GraphQL\Language\AST\OperationDefinitionNode;
 use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Http\Request;
 use Nuwave\Lighthouse\Schema\Context;
@@ -27,11 +29,11 @@ class SubscriberTest extends TestCase
 
         $resolveInfo = $this->createMock(ResolveInfo::class);
         $operationName = 'baz';
-        $resolveInfo->operation = (object) [
-            'name' => (object) [
+        $resolveInfo->operation = new OperationDefinitionNode([
+            'name' => new NameNode([
                 'value' => $operationName,
-            ],
-        ];
+            ]),
+        ]);
         $resolveInfo->fragments = [];
         $context = new Context(new Request());