+++ tests/Unit/Subscriptions/SubscriptionTest.php
@@ -11,7 +11,6 @@ use Nuwave\Lighthouse\Subscriptions\Subscriber;
 use Nuwave\Lighthouse\Subscriptions\SubscriptionBroadcaster;
 use Nuwave\Lighthouse\Subscriptions\SubscriptionRegistry;
 use Nuwave\Lighthouse\Subscriptions\SubscriptionServiceProvider;
-use Prophecy\Argument;
 
 class SubscriptionTest extends SubscriptionTestCase
 {
@@ -26,7 +25,7 @@ class SubscriptionTest extends SubscriptionTestCase
     protected $subscriptionRegistry;
 
     /**
-     * @var \Prophecy\Prophecy\ObjectProphecy
+     * @var \Nuwave\Lighthouse\Subscriptions\SubscriptionBroadcaster&\PHPUnit\Framework\MockObject\MockObject
      */
     protected $broadcaster;
 
@@ -42,17 +41,20 @@ class SubscriptionTest extends SubscriptionTestCase
     {
         parent::setUp();
 
-        $this->schema = "
+        $this->mockResolverExpects($this->any())
+            ->willReturn(self::SUBSCRIPTION_FIELD);
+
+        $this->schema = /** @lang GraphQL */ '
         type Query {
-            subscription: String @field(resolver: \"{$this->qualifyTestResolver()}\")
+            subscription: String @mock
         }
-        ";
+        ';
 
         $this->subscriptionRegistry = app(SubscriptionRegistry::class);
         $this->subscriptionRegistry->register($this->subscription(), self::SUBSCRIPTION_FIELD);
 
-        $this->broadcaster = $this->prophesize(SubscriptionBroadcaster::class);
-        $this->app->instance(BroadcastsSubscriptions::class, $this->broadcaster->reveal());
+        $this->broadcaster = $this->createMock(SubscriptionBroadcaster::class);
+        $this->app->instance(BroadcastsSubscriptions::class, $this->broadcaster);
     }
 
     public function testCanSendSubscriptionToBroadcaster(): void
@@ -63,26 +65,27 @@ class SubscriptionTest extends SubscriptionTestCase
             ],
         ];
 
-        $this->broadcaster->broadcast(
-            Argument::type(GraphQLSubscription::class),
-            self::SUBSCRIPTION_FIELD,
-            $root
-        )->shouldBeCalled();
+        $this->broadcaster
+            ->expects($this->once())
+            ->method('broadcast')
+            ->with(
+                $this->isInstanceOf(GraphQLSubscription::class),
+                self::SUBSCRIPTION_FIELD,
+                $root
+            );
 
         Subscription::broadcast(self::SUBSCRIPTION_FIELD, $root);
     }
 
     public function testThrowsOnInvalidSubscriptionField(): void
     {
-        $this->broadcaster->broadcast(Argument::any())->shouldNotBeCalled();
+        $this->broadcaster
+            ->expects($this->never())
+            ->method('broadcast');
+
         $this->expectException(InvalidArgumentException::class);
 
         Subscription::broadcast('unknownField', []);
-    }
-
-    public function resolve(): string
-    {
-        return self::SUBSCRIPTION_FIELD;
     }
 
     protected function subscription(): GraphQLSubscription